docs(readme): Add project motivation, status badges
* Cleanup
