--- CNAME
@@ -1,1 +1,0 @@
-btph.ga
