+++ README.md
@@ -1,1 +1,19 @@
-## BodyTalks.PH Inventory Management System
+# React Inventory Management System
+
+![yarn-current](https://img.shields.io/badge/Yarn-v1-blue)
+![GitHub package.json dependency version (prod)](https://img.shields.io/github/package-json/dependency-version/decanTyme/practice-spa/react)
+![GitHub package.json dependency version (prod)](https://img.shields.io/github/package-json/dependency-version/decanTyme/practice-spa/@reduxjs/toolkit)
+![GitHub package.json dependency version (prod)](https://img.shields.io/github/package-json/dependency-version/decanTyme/practice-spa/bootstrap)
+![GitHub package.json dependency version (prod)](https://img.shields.io/github/package-json/dependency-version/decanTyme/practice-spa/axios)
+
+## Development
+
+Start the dev server by running:
+
+```sh
+yarn start
+```
+
+## Purpose of Project
+
+A few months ago this project was used as a practice (as the repo name suggests) when I was learning [React](https://reactjs.org/) and [Redux Toolkit](https://redux-toolkit.js.org/), among others. I planned it to be usable enough for our small, humble store but I didn't really have the manpower nor the time to fully implement it at the time. Although this was a particularly ambitious project to create an Inventory Management System (plus Customer Relations) in solo-mode, I have learned _**greatly**_ from it and now I am releasing it publicly. Things have changed I don't think I'll be continuing this anytime soon, especially when I have already found a more feasible meta-framework for solo projects that handles routes much more elegantly, which was the pain-point of the entire system. Some parts of it have already been used in my other projects.
