+++ src/app/pages/common/SpinnerButton.jsx
@@ -4,13 +4,10 @@ import Spinner from "./Spinner";
 function SpinnerButton(props, ref) {
   return (
     <button
-      onClick={props.onClick}
       className={props.className || "btn"}
       data-key={props.itemId}
-      type={props.type}
-      disabled={props.disabled}
-      role={props.role}
       ref={ref}
+      {...props}
     >
       {props.isLoading ? (
         <Spinner addClass="spinner-border-sm">{props.children}</Spinner>
