+++ src/app/pages/components/ThemeManager.jsx
@@ -1,10 +1,19 @@
-import useTheme from "../../../services/hooks/use-theme";
 import { ThemeContext } from "../../../services/providers/theme";
 
 function ThemeProvider({ children }) {
-  const theme = useTheme();
+  const [theme, setTheme] = useState(Constants.DEFAULT_THEME);
+
+  const changeTheme = (theme) =>
+    setTheme(
+      theme === Constants.DEFAULT_THEME
+        ? Constants.THEME_DARK
+        : Constants.THEME_LIGHT
+    );
+
   return (
-    <ThemeContext.Provider value={theme}>{children}</ThemeContext.Provider>
+    <ThemeContext.Provider value={{ theme, changeTheme }}>
+      {children}
+    </ThemeContext.Provider>
   );
 }
